make, and expose to git, development credentials and data
