+++ config.ex.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env_bash
 export SLACK_CLIENT_ID=""
 export SLACK_CLIENT_SECRET=""
-export SLACK_SIGNING_SECRET=""
+export SLACK_SIGNING_SECRET="developmentsigningsecret12345678"
 export SLACK_OAUTH_TOKEN_SECRET=""
 export MONGODB_USERNAME=""
 export MONGODB_PASSWORD=""
